gauss: improve performance of Gauss reduction
Since we expect to encounter many elements which are zero, do not
perform subtraction of a row when the element we attempt to eliminate
is already zero. This makes Gaussian reduction reasonably fast
even for DVB-S2 codes. However, the dense encoder for DVB-S2 codes
uses so much memory that the BER simulation is killed, so the
O(n) "staircase" encoder still needs to be used for DVB-S2.