ber: add simulation of outer BCH decoder
This adds an option to simulate the effects of an outer BCH decoder
that is able to correct up to 't' bit errors, as is used in DVB-S2.

Two separate counts of statistics are maintained. One that looks
at the output of the LDPC decoder only, and one that takes into
account the effects of the outer BCH decoder and considers a codeword
decoded by the LDPC decoder as correct if it has 't' or less bit errors
in the systematic part.

If the BCH decoder simulation is enabled, then the number of codeword
errors at the output of the BCH decoder is used to control when to
finish each simulation step, and the BCH decoder statistics are printed
to the screen in real time. The BCH decoder statistics also get written
to the main output file. There is an additional output file where the
LDPC-only statistics can be written.