add transformation to systematic code
This adds a CLI command that transforms a parity check matrix to
one that can be used for systematic encoding with the first variables
by permuting columns of the parity check matrix.

The crate::encoder::gauss module has been moved to crate::linalg
because it is now used outside crate::encoder too.