update rand to v0.9
Also update other crates related to random numbers. A test for
MacKay-Neal has been modified because it does not pass after updating
these crates. Apparently the underlying implementation for some
rand traits such as choose_multiple has changed, even if the rng
is still reproducible.

This is considered an API-breaking change for ldpc-toolbox.